Move fakeauthorizer code
