+++ coderd/coderd_test.go
@@ -63,7 +63,7 @@ func TestBuildInfo(t *testing.T) {
 // TestAuthorizeAllEndpoints will check `authorize` is called on every endpoint registered.
 func TestAuthorizeAllEndpoints(t *testing.T) {
 	t.Parallel()
-	authorizer := newRecordingAuthorizer()
+	authorizer := &recordingAuthorizer{}
 
 	ctx, cancel := context.WithTimeout(context.Background(), testutil.WaitLong)
 	defer cancel()
@@ -564,28 +564,40 @@ type authCall struct {
 }
 
 type recordingAuthorizer struct {
-	*rbac.FakeAuthorizer
 	Called       *authCall
 	AlwaysReturn error
 }
 
-func newRecordingAuthorizer() *recordingAuthorizer {
-	r := &recordingAuthorizer{}
-	// Use the fake authorizer by rbac to handle prepared authorizers.
-	r.FakeAuthorizer = &rbac.FakeAuthorizer{
-		AuthFunc: func(ctx context.Context, subjectID string, roleNames []string, action rbac.Action, object rbac.Object) error {
-			r.Called = &authCall{
-				SubjectID: subjectID,
-				Roles:     roleNames,
-				Action:    action,
-				Object:    object,
-			}
-			return r.AlwaysReturn
-		},
+func (r *recordingAuthorizer) ByRoleName(_ context.Context, subjectID string, roleNames []string, action rbac.Action, object rbac.Object) error {
+	r.Called = &authCall{
+		SubjectID: subjectID,
+		Roles:     roleNames,
+		Action:    action,
+		Object:    object,
 	}
-	return r
+	return r.AlwaysReturn
+}
+
+func (r *recordingAuthorizer) PrepareByRoleName(_ context.Context, subjectID string, roles []string, action rbac.Action, _ string) (rbac.PreparedAuthorized, error) {
+	return &fakePreparedAuthorizer{
+		Original:  r,
+		SubjectID: subjectID,
+		Roles:     roles,
+		Action:    action,
+	}, nil
 }
 
 func (f *recordingAuthorizer) reset() {
 	f.Called = nil
+}
+
+type fakePreparedAuthorizer struct {
+	Original  *recordingAuthorizer
+	SubjectID string
+	Roles     []string
+	Action    rbac.Action
+}
+
+func (f *fakePreparedAuthorizer) Authorize(ctx context.Context, object rbac.Object) error {
+	return f.Original.ByRoleName(ctx, f.SubjectID, f.Roles, f.Action, object)
 }
